+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Leading Systems Contao Merconis bundle changelog
 ===========================================
 
+### 5.1.16 (2026-06-12)
+ * fix feature 'Widerruf': Configurator/Customizer Referenz
+ * fix branding
+ * fix withdrawal honeypot autofill
+
 ### 5.1.15 (2026-05-08)
  * improve SitemapListener
  * fix default product image livehits